requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
jiter==0.5.0
multidict==6.1.0
openai==1.45.0
pydantic==2.9.1
pydantic_core==2.23.3
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.5
sniffio==1.3.1
telegram==0.0.1
tqdm==4.66.5
typing_extensions==4.12.2
yarl==1.11.1
